WIP Convert readme to Markdown
This allows it to be more or less the same readability in plaintext,
while still rendered nicely and allowing links to headers and such when
used as Markdown, such as on GitHub.

This changeset attempts to retain the existing feel of the document, and
as such the ASCII art logo is untouched and still uses monospace. Sadly,
because starting a line with four spaces indicates a code block, this
must adjust starting whitespace.

In addition to formatting conversion, this:

* Removes trailing whitespace.
* Fixes some spacing, capitalization, and punctuation inconsistencies.